西野さんご指摘修正ver2 9/16
