+++ src/App.tsx
@@ -1,6 +1,6 @@
 import { BrowserRouter, Routes, Route } from 'react-router-dom';
-import Top from './components/Top';
-import Todos from './components/todos';
+import Top from './todos/components/Top';
+import Todos from './todos/components/todos';
 
 function App() {
   return (
